important bugfix
the azimuth value of due west was mistakenly set to 261 degrees. The mistake is thinking that the angle to be used is that which the client requests taking into account the telescope & dome geometry. In fact, dome azimuth is just that. So the azimuth value for west is 270 degrees. Idiot boy...